fix(game.gateway.ts): fix logic for emitting "you're-next" event to the correct socket
The logic for emitting the "you're-next" event to the other connected socket was incorrect. It was checking if the first socket's ID is equal to the current socket's ID, instead of checking if the second socket's ID is equal to the current socket's ID. This has been fixed by changing the index from 0 to 1 in the if condition. Additionally, a console.log statement has been added for debugging purposes.